requirements.txt
gitpython
pandas
PyYAML
numpy